third_party/WebKit/LayoutTests/external/wpt/cssom/getComputedStyle-pseudo-expected.txt

This is a testharness.js-based test.
PASS Resolution of width is correct for ::before and ::after pseudo-elements
PASS Resolution of width is correct for ::before and ::after pseudo-elements of display: contents elements
FAIL Resolution of nonexistent pseudo-element stylesassert_equals: Nonexistent :before pseudo-element shouldn't claim to have definite size expected "auto" but got "784px"
PASS Resolution of pseudo-element styles in display: none elements
Harness: the test ran to completion.
